+++ test/template.test.ts
@@ -1,10 +1,12 @@
 // tslint:disable:no-expression-statement
 import test from 'ava';
+import fs from 'fs-extra';
+import path from 'path';
 import getEngine from '../lib/template';
 
-const templateEngine = getEngine(process.cwd());
+const templateEngine = getEngine(process.cwd(), 'https://example.com/');
 
-test('renderString #1', t => {
+test('clash #1', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `IP-CIDR,67.198.55.0/24,Proxy,no-resolve`;
   const result = templateEngine.renderString(body, {
@@ -14,7 +16,7 @@ test('renderString #1', t => {
   t.is(result, `- IP-CIDR,67.198.55.0/24,Proxy`);
 });
 
-test('renderString #2', t => {
+test('clash #2', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `IP-CIDR,67.198.55.0/24,Proxy,no-resolve // test rule`;
   const result = templateEngine.renderString(body, {
@@ -24,7 +26,7 @@ test('renderString #2', t => {
   t.is(result, `- IP-CIDR,67.198.55.0/24,Proxy`);
 });
 
-test('renderString #3', t => {
+test('clash #3', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `PROCESS-NAME,Telegram,Proxy,no-resolve // test rule`;
   const result = templateEngine.renderString(body, {
@@ -34,7 +36,7 @@ test('renderString #3', t => {
   t.is(result, '');
 });
 
-test('renderString #4', t => {
+test('clash #4', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `# Comment`;
   const result = templateEngine.renderString(body, {
@@ -44,7 +46,7 @@ test('renderString #4', t => {
   t.is(result, '# Comment');
 });
 
-test('renderString #5', t => {
+test('clash #5', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `# Comment`;
   const result = templateEngine.renderString(body, {
@@ -54,7 +56,7 @@ test('renderString #5', t => {
   t.is(result, '# Comment');
 });
 
-test('renderString #6', t => {
+test('clash #6', t => {
   const body = `{{ str | patchYamlArray }}`;
   const str = `URL-REGEX,xxxxxxxxxxxx`;
   const result = templateEngine.renderString(body, {
@@ -62,6 +64,13 @@ test('renderString #6', t => {
   });
 
   t.is(result, '');
+});
+
+test('clash #7', t => {
+  const body = `{{ str | clash }}`;
+
+  t.is(templateEngine.renderString(body, { str: '# test' }), '# test');
+  t.is(templateEngine.renderString(body, { str: '  ' }), '  ');
 });
 
 test('base64', t => {
@@ -75,7 +84,7 @@ test('base64', t => {
   t.is(result, 'dGVzdHRlc3R0ZXN0dGVzdA==');
 });
 
-test('quantumultx filter', t => {
+test('quantumultx filter 1', t => {
   const body = `{{ str | quantumultx }}`;
   const str = `PROCESS-NAME,Telegram,Proxy,no-resolve // test rule`;
   const result = templateEngine.renderString(body, {
@@ -83,6 +92,16 @@ test('quantumultx filter', t => {
   });
 
   t.is(result, '');
+});
+
+test('quantumultx filter 2', t => {
+  const body = `{{ str | quantumultx }}`;
+  const str = fs.readFileSync(path.join(__dirname, './asset/surge-script-list.txt'), { encoding: 'utf8' });
+  const result = templateEngine.renderString(body, {
+    str,
+  });
+
+  t.snapshot(result);
 });
 
 test('mellow filter 1', t => {
@@ -103,4 +122,36 @@ test('mellow filter 2', t => {
   });
 
   t.is(result, '');
+});
+
+test('mellow filter 3', t => {
+  const body = `{{ str | mellow }}`;
+  const str = `# Comment`;
+  const result = templateEngine.renderString(body, {
+    str,
+  });
+
+  t.is(result, '# Comment');
+});
+
+test('spaces in string', t => {
+  const str = `    `;
+
+  t.is(templateEngine.renderString(`{{ str | mellow }}`, { str }), '    ');
+  t.is(templateEngine.renderString(`{{ str | quantumultx }}`, { str }), '    ');
+  t.is(templateEngine.renderString(`{{ str | patchYamlArray }}`, { str }), '    ');
+});
+
+test('ForeignMedia', t => {
+  const str = fs.readFileSync(path.join(__dirname, './asset/ForeignMedia.list'), { encoding: 'utf8' });
+
+  t.snapshot(templateEngine.renderString(`{{ str | quantumultx }}`, {
+    str,
+  }));
+  t.snapshot(templateEngine.renderString(`{{ str | clash }}`, {
+    str,
+  }));
+  t.snapshot(templateEngine.renderString(`{{ str | mellow }}`, {
+    str,
+  }));
 });